Fixed GameLoop pause not accurately tracking game time while paused.  Added FrameSkip button while paused to improve debugging capabilities. Updated BoxCollider to know which direction collision is coming from and improved collision to not go through other colliders. Misc bug fixes.

+++ WindowsApplication/GameScript.h
@@ -15,7 +15,7 @@ namespace FlatEngine
 
 		virtual void Awake() {}
 		virtual void Start() {}
-		virtual void Update(float deltaTime) {}
+		virtual void Update(int deltaTime) {}
 		void SetName(std::string name);
 		std::string GetName();
 		// Entities will be added at game run time for now. Could move this to
